Optymalizacja użycia pamięci

+++ src/komorka.h
@@ -1,8 +1,8 @@
 #ifndef _K_H
 #define _K_H
 typedef struct komorka{
-	int x;
-	int y;
+	short int x;
+	short int y;
 } kmk_t;
 
 #endif
